Organize versioning library
