+++ packages/versioning/lib/main.tsp
@@ -1,0 +1,3 @@
+import "./decorators.tsp";
+import "./projection.tsp";
+import "../dist/src/validate.js";
