+++ packages/versioning/package.json
@@ -18,7 +18,7 @@
   ],
   "type": "module",
   "main": "dist/src/index.js",
-  "tspMain": "lib/versioning.tsp",
+  "tspMain": "lib/main.tsp",
   "exports": {
     ".": {
       "types": "./dist/src/index.d.ts",
