+++ packages/versioning/src/lib.ts
@@ -1,6 +1,10 @@
 import { createTypeSpecLibrary, paramMessage } from "@typespec/compiler";
 
-const libDef = {
+export const {
+  reportDiagnostic,
+  createStateSymbol,
+  stateKeys: VersioningStateKeys,
+} = createTypeSpecLibrary({
   name: "@typespec/versioning",
   diagnostics: {
     "versioned-dependency-tuple": {
@@ -102,5 +106,18 @@ const libDef = {
       },
     },
   },
-} as const;
-export const { reportDiagnostic, createStateSymbol } = createTypeSpecLibrary(libDef);
+  state: {
+    versionIndex: { description: "Version index" },
+
+    addedOn: { description: "State for @addedOn decorator" },
+    removedOn: { description: "State for @removedOn decorator" },
+    versions: { description: "State for @versioned decorator" },
+    useDependencyNamespace: { description: "State for @useDependency decorator on Namespaces" },
+    useDependencyEnum: { description: "State for @useDependency decorator on Enums" },
+    renamedFrom: { description: "State for @renamedFrom decorator" },
+    madeOptional: { description: "State for @madeOptional decorator" },
+    madeRequired: { description: "State for @madeRequired decorator" },
+    typeChangedFrom: { description: "State for @typeChangedFrom decorator" },
+    returnTypeChangedFrom: { description: "State for @returnTypeChangedFrom decorator" },
+  },
+});
