+++ packages/versioning/src/validate.ts
@@ -11,12 +11,8 @@ import {
   Type,
   TypeNameOptions,
 } from "@typespec/compiler";
-import { reportDiagnostic } from "./lib.js";
-import { Version } from "./types.js";
 import {
-  Availability,
   findVersionedNamespace,
-  getAvailabilityMap,
   getMadeOptionalOn,
   getMadeRequiredOn,
   getRenamedFrom,
@@ -24,6 +20,12 @@ import {
   getTypeChangedFrom,
   getUseDependencies,
   getVersion,
+} from "./decorators.js";
+import { reportDiagnostic } from "./lib.js";
+import { Version } from "./types.js";
+import {
+  Availability,
+  getAvailabilityMap,
   getVersionDependencies,
   getVersions,
 } from "./versioning.js";
