+++ packages/versioning/test/library-loading.test.ts
@@ -2,7 +2,7 @@ import { Namespace, projectProgram } from "@typespec/compiler";
 import { BasicTestRunner, createTestWrapper } from "@typespec/compiler/testing";
 import { notStrictEqual, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
-import { buildVersionProjections } from "../src/versioning.js";
+import { buildVersionProjections } from "../src/projection.js";
 import { createVersioningTestHost } from "./test-host.js";
 
 describe("versioning: library loading", () => {
@@ -30,7 +30,7 @@ describe("versioning: library loading", () => {
     };
     // Force loading a different version
     const { buildVersionProjections: buildVersionProjectionsDifferent } = await import(
-      "../src/versioning.js?different=1" as any
+      "../src/projection.js?different=1" as any
     );
     notStrictEqual(
       buildVersionProjections,
