+++ packages/versioning/test/versioned-dependencies.test.ts
@@ -7,7 +7,7 @@ import {
 } from "@typespec/compiler/testing";
 import { ok, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
-import { buildVersionProjections } from "../src/versioning.js";
+import { buildVersionProjections } from "../src/projection.js";
 import { createVersioningTestHost, createVersioningTestRunner } from "./test-host.js";
 import { assertHasProperties } from "./utils.js";
