+++ packages/versioning/test/versioning.test.ts
@@ -20,9 +20,10 @@ import {
 } from "@typespec/compiler/testing";
 import { deepStrictEqual, fail, ok, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
+import { buildVersionProjections, indexTimeline } from "../src/projection.js";
 import { Version } from "../src/types.js";
 import { VersioningTimeline } from "../src/versioning-timeline.js";
-import { buildVersionProjections, getVersions, indexTimeline } from "../src/versioning.js";
+import { getVersions } from "../src/versioning.js";
 import { createVersioningTestHost } from "./test-host.js";
 import {
   assertHasMembers,
